Solve all unpredictability issues ;)
It seems to be very important that we are set to full power (0 PWM
cycle) during the chord change. Otherwise, sometimes when we call
actuate_current_chord(), the 8 bit shift registers will get the wrong
data somehow. This was the original software setup, which I thought
wasn't working originally.

FULL_POWER doesn't have to be on for all that long, only between 25-50
ms, then it can go immediately to LOW_POWER. Keeping this time low seems
to help the power supply/capacitor keep up.

Here's all the possible things that I can think of that may have also
been important

 - I added a big 220uF capacitor across the power source, though
 removing it doesn't seem to resume the hiccups much if at all.
 - I reduced the time at FULL_POWER and made sure FULL_POWER was on
 before actuating chords
 - I set the voltage regulator to between 8 and 9 volts, though it seems
 to work much higher (while generating more noise)
 - Perhaps I set the current limiter on the regulator a little lower? I
 don't know how to quantify this though.