Replace sus2 with dom9
Get rid of sus2 from the keypad entirely. The motivation is that IVsus2
= Isus4, Isus2 = Vsus4, etc. around the circle of 5ths. Furthermore,
after playing around with it there doesn't seem to be a huge auditory
difference.

The choice to move dom9 to the new spot is that I want to try it out
more and it was the only modifier using a double diagonal modifier
combo.