Add a NONE chord combo. Feeble attempt at ALL
The NONE chord is now playable, which just dampens all strings, which is
the same state the instrument starts in. Pretty basic stuff.

The ALL chord is more tricky. Raising every solenoid at once is a huge
power draw, and only sometimes works, pulling the voltage unacceptably
low and sometimes resetting the arduino. Perhaps I should try a simpler
solution; this added function may be too complex. In any case, I
decided not to assign the function on the keypad.