Untested: Make double actuators more "circle-of-fifths-ish"
We currently had the VIb and IIIb as near-the-key triads, playable using
a double modifier combo. Having encountered a need for the IV# (major
sharp IV) I realized that by adding just that and I#, we could have
playable chords built on all 12 semitones, if not flexibility of
tonality for 4 of them.

The right-side double modifiers represent going out the key in the
fourths direction (VIIb -> IIIb -> VIb), and the left-side double
modifiers represent going out of the key in the fifths direction (vii ->
VI# -> I#). Of course, the circle of fifths is a circle, so the choice
to distinguish the two groups of two should be more dependent on how the
chords lead musically, but I don't have a good handle on that yet.

For now, it's just cool to have every (major) triad accessible from any
key!